integrating new endpoints with updated user types and sharing permissions

+++ src/main/java/com/bytecoders/pharmaid/service/PrescriptionService.java
@@ -50,9 +50,10 @@ public class PrescriptionService {
    * @return Prescription the newly created prescription
    */
   public Prescription updatePrescription(Prescription prescription) {
-    // check if user has permissions to create a prescription
+    // check if user has permissions to update a prescription
     permissionValidator.validateEditPermission(jwtUtils.getLoggedInUserId(),
         prescription.getUser().getId());
+
     return prescriptionRepository.save(prescription);
   }
 
@@ -71,7 +72,9 @@ public class PrescriptionService {
    * @param prescriptionId the prescription ID
    */
   public void deletePrescription(String prescriptionId) {
-    prescriptionRepository.deleteById(prescriptionId);
+    if(getPrescription(prescriptionId) != null) {
+      prescriptionRepository.deleteById(prescriptionId);
+    }
   }
 
   /**
